+++ src/cosmos/CosmosEngine.ts
@@ -1,5 +1,9 @@
 import { decodeSignature, encodeSecp256k1Pubkey } from '@cosmjs/amino'
-import { encodePubkey, makeAuthInfoBytes } from '@cosmjs/proto-signing'
+import {
+  EncodeObject,
+  encodePubkey,
+  makeAuthInfoBytes
+} from '@cosmjs/proto-signing'
 import { StargateClient } from '@cosmjs/stargate'
 import { add, gt } from 'biggystring'
 import { SignDoc, TxBody, TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx'
@@ -20,6 +24,7 @@ import { base16 } from 'rfc4648'
 
 import { CurrencyEngine } from '../common/CurrencyEngine'
 import { PluginEnvironment } from '../common/innerPlugin'
+import { MakeTxParams } from '../common/types'
 import { upgradeMemos } from '../common/upgradeMemos'
 import { cleanTxLogs, getFetchCors } from '../common/utils'
 import { CosmosTools } from './CosmosTools'
@@ -29,6 +34,7 @@ import {
   asSafeCosmosWalletInfo,
   asShapeshiftResponse,
   CosmosNetworkInfo,
+  CosmosOtherMethods,
   CosmosWalletOtherData,
   SafeCosmosWalletInfo,
   ShapeshiftTx
@@ -46,6 +52,7 @@ export class CosmosEngine extends CurrencyEngine<
   sequence: number
   fetchCors: EdgeFetchFunction
   otherData!: CosmosWalletOtherData
+  otherMethods: CosmosOtherMethods
 
   constructor(
     env: PluginEnvironment<CosmosNetworkInfo>,
@@ -58,6 +65,50 @@ export class CosmosEngine extends CurrencyEngine<
     this.accountNumber = 0
     this.sequence = 0
     this.fetchCors = getFetchCors(env.io)
+    this.otherMethods = {
+      makeTx: async (params: MakeTxParams) => {
+        switch (params.type) {
+          case 'MakeTxDeposit': {
+            if (this.tools.methods.deposit == null) {
+              throw new Error(
+                `${this.currencyInfo.displayName} does not support the deposit method`
+              )
+            }
+
+            const { assets, memo, metadata } = params
+
+            const msg = this.tools.methods.deposit({
+              assets,
+              memo,
+              signer: this.walletInfo.keys.bech32Address
+            })
+            const otherParams = this.createUnsignedTxHexOtherParams([msg], memo)
+
+            const networkFee = this.networkInfo.defaultTransactionFee.amount
+
+            const out: EdgeTransaction = {
+              blockHeight: 0, // blockHeight,
+              currencyCode: this.currencyInfo.currencyCode,
+              date: Date.now() / 1000,
+              isSend: true,
+              memos: [],
+              metadata,
+              nativeAmount: `-${networkFee}`,
+              networkFee,
+              otherParams,
+              ourReceiveAddresses: [],
+              signedTx: '',
+              txid: '',
+              walletId: this.walletId
+            }
+            return out
+          }
+          default: {
+            throw new Error(`Invalid type: ${params.type}`)
+          }
+        }
+      }
+    }
   }
 
   setOtherData(raw: any): void {
@@ -210,6 +261,20 @@ export class CosmosEngine extends CurrencyEngine<
     this.addTransaction(currencyCode, edgeTransaction)
   }
 
+  private createUnsignedTxHexOtherParams(
+    messages: EncodeObject[],
+    memo?: string
+  ): { unsignedTxHex: string } {
+    const body = TxBody.fromPartial({ messages, memo })
+    const bodyBytes = TxBody.encode(body).finish()
+    const unsignedTxRaw = TxRaw.fromPartial({
+      bodyBytes
+    })
+    return {
+      unsignedTxHex: base16.stringify(TxRaw.encode(unsignedTxRaw).finish())
+    }
+  }
+
   // // ****************************************************************************
   // // Public methods
   // // ****************************************************************************
@@ -265,16 +330,8 @@ export class CosmosEngine extends CurrencyEngine<
       fromAddress: this.walletInfo.keys.bech32Address,
       toAddress: publicAddress
     })
-    const body = TxBody.fromPartial({ messages: [msg], memo })
-    const bodyBytes = TxBody.encode(body).finish()
-    const unsignedTxRaw = TxRaw.fromPartial({
-      bodyBytes
-    })
-    const unsignedTxHex = base16.stringify(TxRaw.encode(unsignedTxRaw).finish())
 
-    const otherParams = {
-      unsignedTxHex
-    }
+    const otherParams = this.createUnsignedTxHexOtherParams([msg], memo)
 
     const edgeTransaction: EdgeTransaction = {
       blockHeight: 0,